requirements.txt
fastapi
python-dotenv
motor
pydantic
uvicorn